new Fields SQL table

+++ models/filtroCodigoProduto.py
@@ -3,4 +3,4 @@ from dataclasses import dataclass
 @dataclass
 class filtroCodigoProduto:
     CODIGO: str
-    QTDE: int
+    QTDE: float
