+++ models/filtroDescricaoProduto.py
@@ -3,4 +3,4 @@ from dataclasses import dataclass
 @dataclass
 class filtroDescricaoProduto:
     DESCRICAO: str
-    QTDE: int
+    QTDE: float
