+++ models/getProduto.py
@@ -2,4 +2,4 @@ from pydantic import BaseModel
 
 class getProduto(BaseModel):
     ID_PRODUTO: int
-    QTDE: int
+    QTDE: float
